Solved Bug (In editor things ran in different order than in build)
In editor current grenades was updtaed before calling player UI and in build was the other way around. So I updated Current grenades before calling the OnLunch greenade action. Code is not as clean as before but it is only one line and it solved the bug. I tried delaying Player UI script execution but it didnt work.